backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.10.1
psycopg[binary]==3.2.10
redis==6.2.0
sqlalchemy==2.0.41
alembic==1.16.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
loguru==0.7.3
email-validator==2.2.0
ruff==0.6.9
mypy==1.11.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
bcrypt==3.2.2

